Add adafruit footprint.
Add Adafruit footprint made from Adafruit PCB model. Fix UDA1334A pin
numbering to match the footprint. This also exposes an issue where the
wsel, bclk, and din pins are not connected to anything because it is
marked as NC on the top level. This should maybe be replaced with male
pin header so dupont connectors can be used to attach these pins to the
raspberry pi.